Move the preview panel test to its own file
